+++ libs/dot-components/src/lib/components/input-form-fields/InputText.tsx
@@ -35,6 +35,7 @@ export const DotInputText = ({
   helperText,
   endIcon,
   id,
+  inputRef,
   label,
   multiline = false,
   name,
@@ -92,6 +93,7 @@ export const DotInputText = ({
           </StyledAdornment>
         ),
       }}
+      inputRef={inputRef}
       label={label}
       multiline={multiline}
       name={name}